feature(ProfanityClient): 비속어 검사 인터페이스 추가
- 기존 구현에 강하게 결합된 PurgomalumClient 네이밍 변경
- FakePurgomalumClient, 가독성 향상
