+++ src/main/java/kitchenpos/products/infra/ProfanityClient.java
@@ -1,0 +1,6 @@
+package kitchenpos.products.infra;
+
+public interface ProfanityClient {
+
+	boolean containsProfanity(String text);
+}
