+++ src/main/java/kitchenpos/products/infra/PurgomalumClient.java
@@ -1,5 +1,27 @@
 package kitchenpos.products.infra;
 
-public interface PurgomalumClient {
-    boolean containsProfanity(String text);
+import java.net.URI;
+
+import org.springframework.boot.web.client.RestTemplateBuilder;
+import org.springframework.stereotype.Component;
+import org.springframework.web.client.RestTemplate;
+import org.springframework.web.util.UriComponentsBuilder;
+
+@Component
+public class PurgomalumClient implements ProfanityClient {
+
+	private final RestTemplate restTemplate;
+
+	public PurgomalumClient(final RestTemplateBuilder restTemplateBuilder) {
+		this.restTemplate = restTemplateBuilder.build();
+	}
+
+	@Override
+	public boolean containsProfanity(final String text) {
+		final URI url = UriComponentsBuilder.fromUriString("https://www.purgomalum.com/service/containsprofanity")
+			.queryParam("text", text)
+			.build()
+			.toUri();
+		return Boolean.parseBoolean(restTemplate.getForObject(url, String.class));
+	}
 }
