+++ src/test/java/kitchenpos/products/application/FakePurgomalumClient.java
@@ -1,20 +1,16 @@
 package kitchenpos.products.application;
 
-import kitchenpos.products.infra.PurgomalumClient;
-
-import java.util.Arrays;
 import java.util.List;
 
-public class FakePurgomalumClient implements PurgomalumClient {
-    private static final List<String> profanities;
+import kitchenpos.products.infra.ProfanityClient;
 
-    static {
-        profanities = Arrays.asList("비속어", "욕설");
-    }
+public class FakePurgomalumClient implements ProfanityClient {
 
-    @Override
-    public boolean containsProfanity(final String text) {
-        return profanities.stream()
-            .anyMatch(profanity -> text.contains(profanity));
-    }
+	private static final List<String> PROFANITIES = List.of("비속어", "욕설");
+
+	@Override
+	public boolean containsProfanity(final String text) {
+		return PROFANITIES.stream()
+			.anyMatch(text::contains);
+	}
 }
